refactor(coral): move browse topics as a subfeature of topics
Resolves: #359
