+++ coral/src/app/features/topics/browse/hooks/environment/useGetEnvironments.test.tsx
@@ -2,7 +2,7 @@ import { QueryClientProvider } from "@tanstack/react-query";
 import { renderHook, waitFor } from "@testing-library/react";
 import { ReactElement } from "react";
 import { server } from "src/services/api-mocks/server";
-import { useGetEnvironments } from "src/app/features/browse-topics/hooks/environment/useGetEnvironments";
+import { useGetEnvironments } from "src/app/features/topics/browse/hooks/environment/useGetEnvironments";
 import { mockGetEnvironments } from "src/domain/environment";
 import { getQueryClientForTests } from "src/services/test-utils/query-client-tests";
 import { mockedEnvironmentResponse } from "src/domain/environment/environment-api.msw";
