+++ coral/src/app/pages/topics/index.tsx
@@ -2,7 +2,7 @@ import { PageHeader } from "@aivenio/aquarium";
 import add from "@aivenio/aquarium/dist/src/icons/add";
 import AuthenticationRequiredBoundary from "src/app/components/AuthenticationRequiredBoundary";
 import PreviewBanner from "src/app/components/PreviewBanner";
-import BrowseTopics from "src/app/features/browse-topics/BrowseTopics";
+import BrowseTopics from "src/app/features/topics/browse/BrowseTopics";
 import Layout from "src/app/layout/Layout";
 
 const Topics = () => {
